more polishes and unused code removal
